change test slurm and slurm fixtures using pytest-docker fixture
